Separate scanner from sequencer
